create simple role for icinga2 and icingaweb2 setup
create role to easily setup icinga2 and icingaweb2 to test it.
If so far everything will be installed on the same host. But can easily be adjusted.
The role it self does not install mariadb or mysql. But I wrote a simple mariadb role for that.
For testing just use the playbook